Change GhostMap_t::CellMask to int because of a weird deadlock on atomics

+++ core/src/amr/AMRmesh.h
@@ -90,7 +90,7 @@ public:
       ZL, ZR,
       FACE_COUNT
     };
-    using CellMask = uint8_t;
+    using CellMask = int;
 
     Kokkos::View<CellMask*> send_cell_masks;
   };
